requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10
itsdangerous==0.24
pymongo==2.8
goose-extractor>=1.0
newspaper==0.0.9.8
lxml==3.6.4